requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
cssselect2==0.7.0
Django==5.0.9
django-cors-headers==4.7.0
django-extensions==3.2.3
django-webpush==0.3.6
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fonttools==4.55.2
frozenlist==1.4.1
google-auth==2.34.0
gunicorn==22.0.0
h11==0.14.0
http_ece==1.2.1
httpcore==1.0.5
httpx==0.27.2
idna==3.9
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.7
onesignal-sdk==2.0.0
packaging==24.0
path==17.0.0
Pillow==10.0.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-vapid==1.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
pyfcm==2.0.7
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.17.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywebpush==2.0.0
redis==5.0.8
reportlab==4.2.5
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
webpush==1.0.2
xlwt==1.3.0
yarl==1.11.1
zopfli==0.2.3.post1
